fix(bot): Gave all permissions to created file, do not fetch if exists

+++ app/bot/bot.py
@@ -40,7 +40,9 @@ class TelegramBot:
 
                 file_path = f'https://api.telegram.org/file/bot{self.token}/{file.file_path}'
 
-                urllib.request.urlretrieve(file_path, f"./../f/images/{file_name}")
+                if not os.path.isfile(f"./../f/images/{file_name}"):
+                    urllib.request.urlretrieve(file_path, f"./../f/images/{file_name}")
+                    os.chmod(file_path, 0o777)
             except:
                 pass
